Add reversible iterator layer for directories
Before this, we would do `.getEntries(...)` every single time a directory iteration would be resumed. This was necessary because the only state WASI can share with us is the position in the directory, and usually we'd know that a file doesn't fit into the buffer only after we'd read it from the iterator.

This caused serious performance issues on large folders due to at least quadratic complexity.

To overcome this problem, we need some way to put an item already read from the iterator, back into it when it doesn't fit into the output buffer. This way, next call to `fd_readdir` would be able to try and read the same item again.

To do this, I've made `getEntries` return a custom iteration object that has one extra `revert(handle)` method. When this method is used, an item is stored in an internal state, and used next time a `next()` is invoked. If there's no such item, a call is proxied to the real underlying directory iterator.

This improved reading of a directory with 2000 files by ~10x.